am263x/am273x: DPL: Fix KW issues and FMEA action items
Fix all the MISRA-C violations for DPL No RTOS files
and execute the action items derived from FMEA

Fixes: MCUSDK-7335, MCUSDK-10784, MCUSDK-10785

+++ source/kernel/nortos/dpl/common/ClockP_nortos.c
@@ -298,9 +298,9 @@ uint64_t ClockP_getTimeUsec(void)
 
 
     /* Get the current time in microseconds */
-	ts = (ticks2 * (uint64_t)gClockCtrl.usecPerTick)
-    +(uint64_t)  ( /* convert timer count to usecs */
-	(uint64_t)((timerCount - gClockCtrl.timerReloadCount)*gClockCtrl.usecPerTick)/((uint64_t)MAX_TIMER_COUNT_VALUE - gClockCtrl.timerReloadCount)
+	ts = ((uint64_t) ticks2 * (uint64_t)gClockCtrl.usecPerTick)
+    +( /* convert timer count to usecs */
+	(((uint64_t)timerCount - (uint64_t)gClockCtrl.timerReloadCount)*(uint64_t)gClockCtrl.usecPerTick)/(uint64_t)(MAX_TIMER_COUNT_VALUE - (uint64_t)gClockCtrl.timerReloadCount)
     );
     return (ts);
 }
